app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.7",
    "langgraph>=0.5.2",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.9",
]
